Fixes #7
We take a view of the numpy arrays used for prediction. This means
we only ask the RNN to predict the next item in the sequence instead
of everything each time.

Not directly comparable to the previous data points because we've
changed splits but it's taking 62% of the time to process 88.3% as
much data.

1766 examples German, 1-layer LSTM, 512 units, generation_timesteps=30

real    3m6.005s
user    2m28.346s
sys     0m36.181s